Moved notifyOnConnectionTimeout to CurrentUser
